# PERF: Improve performance of special attribute lookups