--- a/numpy/core/src/common/get_attr_string.h
+++ b/numpy/core/src/common/get_attr_string.h
@@ -55,7 +55,13 @@ PyArray_LookupSpecial(PyObject *obj, PyObject *name_unicode)
     if (_is_basic_python_type(tp)) {
         return NULL;
     }
-    return PyObject_GetAttr((PyObject *)tp, name_unicode);
+    PyObject * res = PyObject_GetAttr((PyObject *)tp, name_unicode);
+    
+    if (res == NULL && PyErr_ExceptionMatches(PyExc_AttributeError)) {
+        PyErr_Clear();
+    }
+
+    return res;
 }
 
 
@@ -77,7 +83,12 @@ PyArray_LookupSpecial_OnInstance(PyObject *obj, PyObject *name_unicode)
         return NULL;
     }
 
-    return PyObject_GetAttr(obj, name_unicode);
+    PyObject * res = PyObject_GetAttr(obj, name_unicode);
+    if (res == NULL && PyErr_ExceptionMatches(PyExc_AttributeError)) {
+        PyErr_Clear();
+    }
+
+    return res;
 }
 
 #endif  /* NUMPY_CORE_SRC_COMMON_GET_ATTR_STRING_H_ */