This PR reduces the overhead of `ufunc_generic_fastcall`, which is used by many numpy methods. For many input arguments a check is performed on the existence of the `__array_ufunc__` attribute. The check involves a call to `tp_getattro` from `PyTypeObject` which requires a `PyUniCode` object.

This PR avoids construction of the `PyUniCode` for each invocation of the attribute lookup.

Benchmark
```
import pyperf
runner = pyperf.Runner()

runner.timeit(name=f"np.sqrt", stmt=f"np.sqrt(v)", setup='import numpy as np; v=np.float64(1.1)')
runner.timeit(name=f"np.cos", stmt=f"np.cos(v)", setup='import numpy as np; v=np.float64(1.1)')
```
Results
```
np.sqrt: Mean +- std dev: [base] 887 ns +- 37 ns -> [patch] 807 ns +- 49 ns: 1.10x faster
np.cos: Mean +- std dev: [base] 914 ns +- 46 ns -> [patch] 829 ns +- 61 ns: 1.10x faster

Geometric mean: 1.10x faster
```


<details>
<summary>Results of numpy benchmarks </summary>
Benchmark:

```
python runtests.py --bench-compare main bench_ufunc
```
The results show some tests with improved and some with decreased performance. It looks like some instability on my system.
Most changes are in `bench_ufunc_strides.BinaryInt`, on a re-run that values changed and the increases or decreases do not look systematic. 

Full results
```
endebakpt@woelmuis:~/numpy$ python runtests.py --bench-compare main "bench_ufunc"
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.8-Cython
·· Building ee8c683a <performance_cache_unicode_array_ufunc> for virtualenv-py3.8-Cython..................................................
·· Installing ee8c683a <performance_cache_unicode_array_ufunc> into virtualenv-py3.8-Cython.
· Running 68 total benchmarks (2 commits * 1 environments * 34 benchmarks)
[  0.00%] · For numpy commit fd646bd6 <main> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.8-Cython.....................................................
[  0.00%] ·· Benchmarking virtualenv-py3.8-Cython
[  0.00%] ··· Importing benchmark suite produced output:
[  0.00%] ···· NumPy CPU features: SSE SSE2 SSE3 SSSE3* SSE41* POPCNT* SSE42* AVX* F16C* FMA3* AVX2* AVX512F? AVX512CD? AVX512_KNL? AVX512_KNM? AVX512_SKX? AVX512_CLX? AVX512_CNL? AVX512_ICL?
[  0.74%] ··· Running (bench_ufunc.ArgParsing.time_add_arg_parsing--).........................
[ 19.12%] ··· Running (bench_ufunc_strides.AVX_UFunc_log.time_log--).....
[ 22.79%] ··· Running (bench_ufunc_strides.BinaryInt.time_ufunc--).
[ 23.53%] ··· Running (bench_ufunc_strides.LogisticRegression.time_train--)..
[ 25.00%] ··· Running (bench_ufunc_strides.Unary.time_ufunc--).
[ 25.00%] · For numpy commit ee8c683a <performance_cache_unicode_array_ufunc> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.8-Cython..
[ 25.00%] ·· Benchmarking virtualenv-py3.8-Cython
[ 25.74%] ··· Running (bench_ufunc.ArgParsing.time_add_arg_parsing--).........................
[ 44.12%] ··· Running (bench_ufunc_strides.AVX_UFunc_log.time_log--).....
[ 47.79%] ··· Running (bench_ufunc_strides.BinaryInt.time_ufunc--).
[ 48.53%] ··· Running (bench_ufunc_strides.LogisticRegression.time_train--)..
[ 50.00%] ··· Running (bench_ufunc_strides.Unary.time_ufunc--).
[ 50.00%] · For numpy commit ee8c683a <performance_cache_unicode_array_ufunc> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.8-Cython
[ 50.74%] ··· bench_ufunc.ArgParsing.time_add_arg_parsing                                                                                                                                  ok
[ 50.74%] ··· =============================================================== ==========
                                         arg_kwarg                                      
              --------------------------------------------------------------- ----------
                                   (array(1.), array(2.))                      710±60ns 
                             (array(1.), array(2.), array(3.))                 598±10ns 
                           (array(1.), array(2.), out=array(3.))               681±20ns 
                          (array(1.), array(2.), out=(array(3.),))             660±30ns 
               (array(1.), array(2.), out=array(3.), subok=True, where=True)   692±20ns 
                             (array(1.), array(2.), subok=True)                788±60ns 
                       (array(1.), array(2.), subok=True, where=True)          823±60ns 
                 (array(1.), array(2.), array(3.), subok=True, where=True)     695±30ns 
              =============================================================== ==========

[ 51.47%] ··· bench_ufunc.ArgParsingReduce.time_add_reduce_arg_parsing                                                                                                                     ok
[ 51.47%] ··· ====================================================== =============
                                    arg_kwarg                                     
              ------------------------------------------------------ -------------
                                (array([0., 1.]))                     1.34±0.01μs 
                               (array([0., 1.]), 0)                   1.36±0.02μs 
                            (array([0., 1.]), axis=0)                 1.42±0.02μs 
                            (array([0., 1.]), 0, None)                1.36±0.02μs 
                      (array([0., 1.]), axis=0, dtype=None)           1.44±0.01μs 
                      (array([0., 1.]), 0, None, array(0.))             1.23±0μs  
               (array([0., 1.]), axis=0, dtype=None, out=array(0.))   1.29±0.01μs 
                         (array([0., 1.]), out=array(0.))             1.28±0.03μs 
              ====================================================== =============

[ 52.21%] ··· bench_ufunc.Broadcast.time_broadcast                                                                                                                                10.5±0.04ms
[ 52.94%] ··· bench_ufunc.Custom.time_and_bool                                                                                                                                    1.68±0.02μs
[ 53.68%] ··· bench_ufunc.Custom.time_nonzero                                                                                                                                      13.1±0.1μs
[ 54.41%] ··· bench_ufunc.Custom.time_not_bool                                                                                                                                    1.46±0.03μs
[ 55.15%] ··· bench_ufunc.Custom.time_or_bool                                                                                                                                     1.61±0.02μs
[ 55.88%] ··· bench_ufunc.CustomArrayFloorDivideInt.time_floor_divide_int                                                                                                                  ok
[ 55.88%] ··· ============== ============= ============= =============
              --                                size                  
              -------------- -----------------------------------------
                  dtype           100          10000        1000000   
              ============== ============= ============= =============
                numpy.int8    1.11±0.01μs    73.7±0.3μs   7.87±0.03ms 
               numpy.int16    1.11±0.02μs    75.1±0.3μs   8.29±0.05ms 
               numpy.int32    1.07±0.03μs    94.0±0.6μs   10.3±0.05ms 
               numpy.int64    1.64±0.02μs     161±1μs      17.6±0.2ms 
               numpy.uint8    1.00±0.03μs    38.7±0.1μs   3.76±0.04ms 
               numpy.uint16   1.02±0.03μs   38.8±0.06μs   3.79±0.01ms 
               numpy.uint32   1.03±0.03μs    38.7±0.1μs   3.83±0.04ms 
               numpy.uint64   1.44±0.03μs    82.3±0.3μs   8.27±0.06ms 
              ============== ============= ============= =============

[ 56.62%] ··· bench_ufunc.CustomInplace.time_char_or                                                                                                                               46.4±0.2μs
[ 57.35%] ··· bench_ufunc.CustomInplace.time_char_or_temp                                                                                                                          60.0±0.5μs
[ 58.09%] ··· bench_ufunc.CustomInplace.time_double_add                                                                                                                            56.6±0.1μs
[ 58.82%] ··· bench_ufunc.CustomInplace.time_double_add_temp                                                                                                                       75.1±0.1μs
[ 59.56%] ··· bench_ufunc.CustomInplace.time_float_add                                                                                                                             57.5±0.3μs
[ 60.29%] ··· bench_ufunc.CustomInplace.time_float_add_temp                                                                                                                        76.1±0.3μs
[ 61.03%] ··· bench_ufunc.CustomInplace.time_int_or                                                                                                                                56.5±0.2μs
[ 61.76%] ··· bench_ufunc.CustomInplace.time_int_or_temp                                                                                                                           71.6±0.6μs
[ 62.50%] ··· bench_ufunc.CustomScalar.time_add_scalar2                                                                                                                                    ok
[ 62.50%] ··· =============== ============
                   dtype                  
              --------------- ------------
               numpy.float32   6.17±0.2μs 
               numpy.float64   12.8±0.1μs 
              =============== ============

[ 63.24%] ··· bench_ufunc.CustomScalar.time_divide_scalar2                                                                                                                                 ok
[ 63.24%] ··· =============== =============
                   dtype                   
              --------------- -------------
               numpy.float32   12.4±0.06μs 
               numpy.float64    26.4±0.2μs 
              =============== =============

[ 63.97%] ··· bench_ufunc.CustomScalar.time_divide_scalar2_inplace                                                                                                                         ok
[ 63.97%] ··· =============== ============
                   dtype                  
              --------------- ------------
               numpy.float32   12.7±0.2μs 
               numpy.float64   26.4±0.3μs 
              =============== ============

[ 64.71%] ··· bench_ufunc.CustomScalar.time_less_than_scalar2                                                                                                                              ok
[ 64.71%] ··· =============== =============
                   dtype                   
              --------------- -------------
               numpy.float32   4.05±0.08μs 
               numpy.float64    7.19±0.1μs 
              =============== =============

[ 65.44%] ··· bench_ufunc.CustomScalarFloorDivideInt.time_floor_divide_int                                                                                                                 ok
[ 65.44%] ··· ============== ============= ============= ============= =============
              --                                     divisors                       
              -------------- -------------------------------------------------------
                  dtype            8             -8            43           -43     
              ============== ============= ============= ============= =============
                numpy.int8    3.03±0.08μs    3.16±0.2μs   3.01±0.09μs    3.08±0.2μs 
               numpy.int16    3.32±0.07μs   3.23±0.08μs    3.36±0.2μs    3.34±0.1μs 
               numpy.int32    5.60±0.02μs    5.51±0.1μs   5.61±0.05μs   5.44±0.06μs 
               numpy.int64    14.9±0.03μs    15.1±0.1μs    14.9±0.1μs   14.8±0.06μs 
               numpy.uint8     3.24±0.1μs       n/a       3.24±0.07μs       n/a     
               numpy.uint16    3.63±0.1μs       n/a       3.70±0.08μs       n/a     
               numpy.uint32    5.40±0.2μs       n/a        5.43±0.1μs       n/a     
               numpy.uint64    12.1±0.1μs       n/a        12.1±0.2μs       n/a     
              ============== ============= ============= ============= =============

[ 65.44%] ···· For parameters: <class 'numpy.uint8'>, -8
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint8'>, -43
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint16'>, -8
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint16'>, -43
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint32'>, -8
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint32'>, -43
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint64'>, -8
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')
               
               For parameters: <class 'numpy.uint64'>, -43
               asv: skipped: NotImplementedError('Skipping test for negative divisor with unsigned type')

[ 66.18%] ··· bench_ufunc.Scalar.time_add_scalar                                                                                                                                     575±50ns
[ 66.91%] ··· bench_ufunc.Scalar.time_add_scalar_conv                                                                                                                                803±20ns
[ 67.65%] ··· bench_ufunc.Scalar.time_add_scalar_conv_complex                                                                                                                        825±30ns
[ 68.38%] ··· bench_ufunc.UFunc.time_ufunc_types                                                                                                                                           ok
[ 68.38%] ··· =============== =============
                   ufunc                   
              --------------- -------------
                    abs          808±4μs   
                  absolute       813±4μs   
                    add          385±1μs   
                   arccos      6.26±0.04ms 
                  arccosh      5.93±0.02ms 
                   arcsin      6.32±0.06ms 
                  arcsinh      5.82±0.05ms 
                   arctan      3.47±0.01ms 
                  arctan2        1.84±0ms  
                  arctanh      3.73±0.02ms 
                bitwise_and     32.8±0.3μs 
                bitwise_not    21.2±0.04μs 
                 bitwise_or     32.6±0.1μs 
                bitwise_xor     32.6±0.1μs 
                    cbrt       1.85±0.01ms 
                    ceil         208±1μs   
                    conj        189±0.6μs  
                 conjugate       190±1μs   
                  copysign      177±0.7μs  
                    cos        6.95±0.03ms 
                    cosh       6.12±0.05ms 
                  deg2rad        240±1μs   
                  degrees       240±0.9μs  
                   divide        681±4μs   
                   divmod        1.01±0ms  
                   equal         301±3μs   
                    exp        4.90±0.05ms 
                    exp2       4.70±0.03ms 
                   expm1       9.31±0.04ms 
                    fabs        237±0.8μs  
                float_power    10.3±0.04ms 
                   floor        209±0.4μs  
                floor_divide     887±2μs   
                    fmax        440±0.8μs  
                    fmin        434±0.5μs  
                    fmod         591±2μs   
                   frexp         342±2μs   
                    gcd         215±0.5μs  
                  greater        307±1μs   
               greater_equal     301±1μs   
                 heaviside       387±2μs   
                   hypot       1.04±0.01ms 
                   invert       21.5±0.3μs 
                  isfinite      169±0.7μs  
                   isinf        170±0.5μs  
                   isnan        146±0.3μs  
                   isnat         336±2ns   
                    lcm         336±0.9μs  
                   ldexp         213±1μs   
                 left_shift     98.8±0.3μs 
                    less        296±0.7μs  
                 less_equal      292±2μs   
                    log        3.12±0.01ms 
                   log10       3.35±0.02ms 
                   log1p       3.32±0.02ms 
                    log2       3.20±0.03ms 
                 logaddexp       348±10μs  
                 logaddexp2      330±1μs   
                logical_and      317±5μs   
                logical_not      200±2μs   
                 logical_or     262±0.8μs  
                logical_xor      379±6μs   
                   matmul       22.7±0.3ms 
                  maximum        416±2μs   
                  minimum        397±1μs   
                    mod          697±2μs   
                    modf         435±2μs   
                  multiply       397±1μs   
                  negative       224±1μs   
                 nextafter       399±2μs   
                 not_equal       311±2μs   
                  positive       231±1μs   
                   power       10.8±0.02ms 
                  rad2deg        240±1μs   
                  radians       240±0.7μs  
                 reciprocal      724±1μs   
                 remainder       697±3μs   
                right_shift     102±0.4μs  
                    rint         417±4μs   
                    sign         249±3μs   
                  signbit       92.1±0.4μs 
                    sin        6.70±0.05ms 
                    sinh       6.64±0.04ms 
                  spacing        432±3μs   
                    sqrt       1.52±0.01ms 
                   square        244±3μs   
                  subtract       379±2μs   
                    tan        8.30±0.04ms 
                    tanh       5.79±0.02ms 
                true_divide      702±40μs  
                   trunc        204±0.7μs  
              =============== =============

[ 69.12%] ··· bench_ufunc_strides.AVX_UFunc_log.time_log                                                                                                                                   ok
[ 69.12%] ··· ======== ============= ============
              --                 dtype           
              -------- --------------------------
               stride        f            d      
              ======== ============= ============
                 1       22.1±0.8μs   59.9±0.4μs 
                 2      28.8±0.09μs   60.4±0.4μs 
                 4       29.9±0.2μs   60.9±0.2μs 
              ======== ============= ============

…