@seberg 2022-05-02
Oh, I had not realized that this branch kicks in for our scalars.  I think this makes sense in any case, although I am curious about cleaning the code up a bit more.  Would you be up for that?

The main point is that I am pretty sure there are two things to note here:
1. `maybe_get_attr` has probably no meaningful advantage over `PyObject_GetAttr` when we already have the strings (at least not anymore).  The reason is twofold:
   * The comment reads incorrect.  The function suppresses exceptions, it only avoids them if an object has no attributes at all.  But that should be no meaningful object in existence.
   * The `char *` version will be cached on the unicode object.
   I suppose our own version might avoid a tiny bit of extra indirections, etc. but I got the feeling that it is probably not meaningful.
2. This is internal API, that is used only in about 10 places.  It would be awesome to just change it everywhere to avoid duplicated code.

There are two ways to go about, either introduce a `static` in each function to cache the interned string, or add it [here](https://github.com/numpy/numpy/blob/ecd32626dee3cce61ccb91510f61dcd8be8b2122/numpy/core/src/umath/umathmodule.c#L222) and/or [here](https://github.com/numpy/numpy/blob/ecd32626dee3cce61ccb91510f61dcd8be8b2122/numpy/core/src/multiarray/multiarraymodule.c#L4677) as necessary.  Either seems fine to me.

I don't want to overstretch you though.  So if you like, I could also try to do the larger refactor and you can review it instead?