@eendebakpt 2022-05-02
@seberg I can take a shot a refactoring. There is an issue though with both `PyObject_GetAttr` and the current `maybe_get_attr` implementation. For the numpy scalars there is no attribute `__array_ufunc__`, which means that `PyObject_GetAttr` generates an exception that has to be cleared later (`PyObject_HasAttr` combines that). Generating and clearing the exception is quite expensive.

A screenshot of profiling results from calculating `np.sqrt(np.float64(1.1))`:

![sqrt-profile](https://user-images.githubusercontent.com/883786/166313754-73927b3e-1422-43d3-8656-914f95a177ea.png)

About 37% of the `ufunc_generic_fastcall` is spend in `PyUFync_getNonDefaultArrayUFunc`. About 10% is in `PyUnicode_InternFromString` (this is resolved by this PR), and about 22% in generating a formatted exception for missing the attribute `__array_ufunc__`. The methods `PyObject_GetAttr` and PyObject_HasAttr` both suffer from this as well.

There are some options to eliminate the exception generation:

1. Add a fast return for the numpy scalar types (e.g. `np.float64`) just as there is for `np.ndarray` See https://github.com/numpy/numpy/blob/main/numpy/core/src/common/ufunc_override.c#L30 

2. a. Implement a specialized version of `PyObject_GetAttr` without the exception generation in numpy
    b. The same, but try to implement it upstream in cpython
3. Set `__array_ufunc__` to `None` on numpy scalars as getting an existing attribute is much faster than finding out an attribute does not exist. This does require numpy being able to handle the case where `__array_ufunc__` is `None` 

Since I an new to numpy I cannot see the implications of these options (and perhaps there are more), so I welcome your thoughts on this.