@eendebakpt 2022-05-02
@seberg The attribute lookup is slow due to the fact that numpy looks for the attribute on the type and not the instance
```
/*
 * Lookup a special method, following the python approach of looking up
 * on the type object, rather than on the instance itself.
 *
 * Assumes that the special method is a numpy-specific one, so does not look
 * at builtin types, nor does it look at a base ndarray.
 *
 * In future, could be made more like _Py_LookupSpecial
 */
static NPY_INLINE PyObject *
PyArray_LookupSpecial(PyObject *obj, char const *name)
...
```
The following benchmarks shows what happens:
```
import numpy as np
x=np.float64(1.1)

%timeit hasattr(x, '__array_ufunc__')
%timeit hasattr(x, 'all')
%timeit hasattr(np.float64, 'all')
%timeit hasattr(np.float64, '__array_ufunc__') 
```
Result:
```
53.9 ns ± 0.65 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
67.2 ns ± 0.488 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
93.3 ns ± 0.251 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
268 ns ± 0.569 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
The difference in execution times is because in the first 2 tests we follow this path:

https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/object.c#L952

and in the last two we follow

1. https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/object.c#L961
2. https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/typeobject.c#L4411
3. https://github.com/python/cpython/blob/364ed9409269fb321dc4eafdea677c09a4bc0d8d/Objects/typeobject.c#L3902

Not sure why the attribute lookup for types should be slower, I will look into this.