@seberg 2022-05-02
About the earlier option:
1. A fast return may work. It does need to check the super class, so it is not nearly as fast as an "exact" check, though.  But since we likely got an arbitrary object that may be OK at that point.
2. Not sure a special GetAttr is possible?  Since `getattr` is a method?
3. `None` already has the meaning of "ufuncs" not supported, it should work to set it to the same value as `ndarray.__array_ufunc__` or another singleton though.

Looking up on the type is correct, I am surprised that looking up on the instance is so much faster!  Squinting at the Python code, it might be to do with metaclass handling, but these don't even have a metaclass (well besides `type` itself).
I guess there might be some way to improve that (in Python or NumPy), but I need to look at it for longer.

---

But, no matter all of those other things, storing the interned string somewhere and working with that exclusively still seems like a good start, that is independend of the other things you found?