@eendebakpt 2022-05-03
The third option works:
```
import numpy as np
from numpy import sin, cos, sqrt
import time

class myfloat(np.float64):
    pass
myfloat.__array_ufunc__ = np.ndarray.__array_ufunc__


def test(my_class):
    for jj in range(1000):
        v=my_class(np.random.rand())    
        for kk in range(2000):
            w=sqrt(v) + cos(v) + sin(v)

t0=time.perf_counter()
test(np.float64)
dt1=time.perf_counter()-t0
print(dt1)

t0=time.perf_counter()
test(myfloat)
dt2=time.perf_counter()-t0
print(dt2)

print(f'factor {dt1/dt2:.2f} faster')
```
results in a 5% speedup (and perhaps when added on the C side). I will first complete refactoring the unicode string conversion and then continue on the attribute lookup.