@seberg 2022-05-03
Yeah, but I had not double checked that it will not have any effect on binary operators in weird subclassing situation.  Not that this is likely to matter in practice :/.  Defining `__array_ufunc__` would tell NumPy that `x + myobj` does not need to call `myobj.__radd__` unfortunately.  So we would need to introduce a new singleton to get that right  (or say it is OK with a release note, maybe.)

The alternative would be to use `PyArray_CheckAnyScalarExact` which may well be fast enough in practice.

Squinting at it, if it is important to not slow down unknown/arbitrary objects in some of these places, we could even add a bloom filter (a bloom filter is like a mini `set` that can only tell you that something *may* be included).  But I doubt that is worthwhile, more of a fun thought ;).
It might for example make sense for "May be a Python builtin object", only then we need to check them all.

Lets focus on the other things first!  Interesting CPython issue, so it is the error formatting mainly/only?  I guess CPython would have to add a fast-path for `type` there to avoid the error...